Rename a network and fixed activation functions
